+++ resource/resource_carShare.go
@@ -9,6 +9,7 @@ import (
 	"github.com/LewisWatson/carshare-back/model"
 	"github.com/LewisWatson/carshare-back/storage"
 	"github.com/manyminds/api2go"
+	"gopkg.in/LewisWatson/firebase-jwt-auth.v1"
 )
 
 // CarShareResource for api2go routes
@@ -16,18 +17,30 @@ type CarShareResource struct {
 	CarShareStorage storage.CarShareStorage
 	TripStorage     storage.TripStorage
 	UserStorage     storage.UserStorage
+	TokenVerifier   fireauth.TokenVerifier
 }
 
 // FindAll to satisfy api2go.FindAll interface
 func (cs CarShareResource) FindAll(r api2go.Request) (api2go.Responder, error) {
-	result, err := cs.CarShareStorage.GetAll(r.Context)
+
+	userID, err := verify(r, cs.TokenVerifier, cs.UserStorage)
 	if err != nil {
 		return &Response{}, api2go.NewHTTPError(
-			fmt.Errorf("Error retrieveing all car shares, %s", err),
+			err,
+			http.StatusText(http.StatusForbidden),
+			http.StatusForbidden,
+		)
+	}
+
+	result, err := cs.CarShareStorage.GetAll(userID, r.Context)
+	if err != nil {
+		return &Response{}, api2go.NewHTTPError(
+			fmt.Errorf("Error retrieving all car shares, %s", err),
 			http.StatusText(http.StatusInternalServerError),
 			http.StatusInternalServerError,
 		)
 	}
+
 	/*
 	 * Populate the car share relationships. If an error occurs while we do this
 	 * then return the error along with what has been retrieved up to that point
@@ -49,6 +62,16 @@ func (cs CarShareResource) FindAll(r api2go.Request) (api2go.Responder, error) {
 
 // FindOne to satisfy api2go.CRUD interface
 func (cs CarShareResource) FindOne(ID string, r api2go.Request) (api2go.Responder, error) {
+
+	userID, err := verify(r, cs.TokenVerifier, cs.UserStorage)
+	if err != nil {
+		return &Response{}, api2go.NewHTTPError(
+			fmt.Errorf("Error retrieving car share, %s", err),
+			http.StatusText(http.StatusForbidden),
+			http.StatusForbidden,
+		)
+	}
+
 	carShare, err := cs.CarShareStorage.GetOne(ID, r.Context)
 	switch err {
 	case nil:
@@ -67,6 +90,15 @@ func (cs CarShareResource) FindOne(ID string, r api2go.Request) (api2go.Responde
 			http.StatusInternalServerError,
 		)
 	}
+
+	if !carShare.IsMember(userID) {
+		return &Response{}, api2go.NewHTTPError(
+			fmt.Errorf("User %v not member of car share %v", userID, carShare.GetID()),
+			http.StatusText(http.StatusForbidden),
+			http.StatusForbidden,
+		)
+	}
+
 	// if an error occurs while populating, still attempt to send the remainder
 	// of the response
 	popErr := cs.populate(&carShare, r.Context)
@@ -83,6 +115,16 @@ func (cs CarShareResource) FindOne(ID string, r api2go.Request) (api2go.Responde
 
 // Create to satisfy api2go.CRUD interface
 func (cs CarShareResource) Create(obj interface{}, r api2go.Request) (api2go.Responder, error) {
+
+	userID, err := verify(r, cs.TokenVerifier, cs.UserStorage)
+	if err != nil {
+		return &Response{}, api2go.NewHTTPError(
+			fmt.Errorf("Error creating car shares, %s", err),
+			http.StatusText(http.StatusForbidden),
+			http.StatusForbidden,
+		)
+	}
+
 	carShare, ok := obj.(model.CarShare)
 	if !ok {
 		return &Response{}, api2go.NewHTTPError(
@@ -91,10 +133,20 @@ func (cs CarShareResource) Create(obj interface{}, r api2go.Request) (api2go.Res
 			http.StatusBadRequest,
 		)
 	}
+
+	if !carShare.IsMember(userID) {
+		carShare.MemberIDs = append(carShare.MemberIDs, userID)
+	}
+
+	if !carShare.IsAdmin(userID) {
+		carShare.AdminIDs = append(carShare.AdminIDs, userID)
+	}
+
 	id, err := cs.CarShareStorage.Insert(carShare, r.Context)
 	if err == nil && id == "" {
 		err = errors.New("null id returned")
 	}
+
 	if err != nil {
 		errMsg := "Error occurred while persisting car share"
 		return &Response{}, api2go.NewHTTPError(
@@ -103,7 +155,9 @@ func (cs CarShareResource) Create(obj interface{}, r api2go.Request) (api2go.Res
 			http.StatusInternalServerError,
 		)
 	}
+
 	carShare.SetID(id)
+
 	// if an error occurs while populating, still attempt to send the remainder
 	// of the response
 	popErr := cs.populate(&carShare, r.Context)
@@ -115,11 +169,21 @@ func (cs CarShareResource) Create(obj interface{}, r api2go.Request) (api2go.Res
 			http.StatusInternalServerError,
 		)
 	}
+
 	return &Response{Res: carShare, Code: http.StatusCreated}, err
 }
 
 // Delete to satisfy api2go.CRUD interface
 func (cs CarShareResource) Delete(id string, r api2go.Request) (api2go.Responder, error) {
+
+	userID, err := verify(r, cs.TokenVerifier, cs.UserStorage)
+	if err != nil {
+		return &Response{}, api2go.NewHTTPError(
+			fmt.Errorf("Error deleting car share, %s", err),
+			http.StatusText(http.StatusForbidden),
+			http.StatusForbidden,
+		)
+	}
 
 	carShare, err := cs.CarShareStorage.GetOne(id, r.Context)
 	switch err {
@@ -137,6 +201,14 @@ func (cs CarShareResource) Delete(id string, r api2go.Request) (api2go.Responder
 			fmt.Errorf("%s, %s", errMsg, err),
 			errMsg,
 			http.StatusInternalServerError,
+		)
+	}
+
+	if !carShare.IsAdmin(userID) {
+		return &Response{}, api2go.NewHTTPError(
+			fmt.Errorf("Non admin user %v attempting to delete car share %v", userID, carShare.GetID()),
+			http.StatusText(http.StatusForbidden),
+			http.StatusForbidden,
 		)
 	}
 
@@ -159,16 +231,8 @@ func (cs CarShareResource) Delete(id string, r api2go.Request) (api2go.Responder
 		)
 	}
 
-	var tripsDeletedOK = true
-	for _, tripID := range carShare.TripIDs {
-		err = cs.TripStorage.Delete(tripID, r.Context)
-		if err != nil && err != storage.ErrNotFound {
-			tripsDeletedOK = false
-			log.Printf("Error deleting associated trip %s, %v", tripID, err)
-		}
-	}
-
-	if !tripsDeletedOK {
+	ok := cs.deleteAssocTrips(carShare, r.Context)
+	if !ok {
 		errMsg := fmt.Sprintf("Car share deleted, but error occurred while deleting associated trips")
 		return &Response{}, api2go.NewHTTPError(
 			fmt.Errorf("%s", errMsg),
@@ -180,8 +244,29 @@ func (cs CarShareResource) Delete(id string, r api2go.Request) (api2go.Responder
 	return &Response{Code: http.StatusOK}, nil
 }
 
+func (cs CarShareResource) deleteAssocTrips(carShare model.CarShare, ctx api2go.APIContexter) bool {
+	ok := true
+	for _, tripID := range carShare.TripIDs {
+		err := cs.TripStorage.Delete(tripID, ctx)
+		if err != nil && err != storage.ErrNotFound {
+			ok = false
+			log.Printf("Error deleting associated trip %s, %v", tripID, err)
+		}
+	}
+	return ok
+}
+
 // Update to satisfy api2go.CRUD interface
 func (cs CarShareResource) Update(obj interface{}, r api2go.Request) (api2go.Responder, error) {
+
+	userID, err := verify(r, cs.TokenVerifier, cs.UserStorage)
+	if err != nil {
+		return &Response{}, api2go.NewHTTPError(
+			fmt.Errorf("Error updating car share, %s", err),
+			http.StatusText(http.StatusForbidden),
+			http.StatusForbidden,
+		)
+	}
 
 	carShare, ok := obj.(model.CarShare)
 	if !ok {
@@ -189,6 +274,23 @@ func (cs CarShareResource) Update(obj interface{}, r api2go.Request) (api2go.Res
 			fmt.Errorf("Invalid instance given to car share update: %v", obj),
 			http.StatusText(http.StatusBadRequest),
 			http.StatusBadRequest,
+		)
+	}
+
+	existingCarShare, err := cs.CarShareStorage.GetOne(carShare.GetID(), r.Context)
+	if err != nil {
+		return &Response{}, api2go.NewHTTPError(
+			fmt.Errorf("Unable to find carShare %v", carShare.GetID()),
+			http.StatusText(http.StatusNotFound),
+			http.StatusNotFound,
+		)
+	}
+
+	if !existingCarShare.IsAdmin(userID) {
+		return &Response{}, api2go.NewHTTPError(
+			fmt.Errorf("Non admin user %v attempting to update carShare %v", userID, carShare.GetID()),
+			http.StatusText(http.StatusForbidden),
+			http.StatusForbidden,
 		)
 	}
 
@@ -244,7 +346,7 @@ func (cs CarShareResource) Update(obj interface{}, r api2go.Request) (api2go.Res
 		}
 	}
 
-	err := cs.CarShareStorage.Update(carShare, r.Context)
+	err = cs.CarShareStorage.Update(carShare, r.Context)
 	switch err {
 	case nil:
 		break