+++ resource/resource_suite_test.go
@@ -5,6 +5,7 @@ import (
 
 	. "github.com/onsi/ginkgo"
 	. "github.com/onsi/gomega"
+	"gopkg.in/jose.v1/jwt"
 	mgo "gopkg.in/mgo.v2"
 	dockertest "gopkg.in/ory-am/dockertest.v3"
 
@@ -36,3 +37,12 @@ var _ = AfterSuite(func() {
 		}
 	}
 })
+
+type mockTokenVerifier struct {
+	Claims jwt.Claims
+	Error  error
+}
+
+func (mtv mockTokenVerifier) Verify(accessToken string) (userID string, claims jwt.Claims, err error) {
+	return mtv.Claims.Get("sub").(string), mtv.Claims, mtv.Error
+}