+++ CHANGELOG.md
@@ -1,4 +1,5 @@
 # Change Log
+
 All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](http://keepachangelog.com/)
@@ -7,20 +8,26 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 ## [Unreleased]
 
 ## Added
- - CORS headers
+
+- CORS headers
+- Firebase JWT Validation
+- Restrict access by user
 
 ## Fixed
- - Relation links between car shares, trips and users
+
+- Relation links between car shares, trips and users
 
 ## [0.2.0] - 2016-12-10
 
 ### Added
+
 - MongoDB data store support
 - Configure MongoDB URL via `CARSHARE_MGO_URL` environment variable
 - Configure server port via `CARSHARE_PORT` environment variable
 - Created Dockerfile
 
 ### Changed
+
 - Standardisd ID's on [BSON ObjectId](https://docs.mongodb.com/manual/reference/bson-types/#objectid)
 - Overhaul error handling
 - Unit test now run twice. First with the fast in-memory data store (fail fast), then as an integration test against a MongoDB docker container